GooseShop site V - 16/06/2023 - local

+++ wp-content/plugins/blocksy-companion/framework/features/demo-install/content-eraser.php
@@ -91,6 +91,7 @@ class DemoInstallContentEraser {
 			$term = get_term($term_id);
 
 			if (! $term) continue;
+			if (is_wp_error($term)) continue;
 
 			wp_delete_term($term_id, $term->taxonomy);
 		}
